updating stuff for importer and speed over time

+++ ATSPM/Data/Models/SpeedManagement/MonthlyAggregation/MonthlyAggregation.cs
@@ -6,7 +6,7 @@ public class MonthlyAggregation
     public DateTime? CreatedDate { get; set; } //This is the timestamp of when this entry was created.
     public DateTime BinStartTime { get; set; } // This is the 
     public Guid SegmentId { get; set; }
-    public int SourceId { get; set; }
+    public long SourceId { get; set; }
     public long? AllDayAverageSpeed { get; set; } //All Day 12:00AM - 11:59PM
     public long? AllDayViolations { get; set; }
     public long? AllDayExtremeViolations { get; set; }
